+++ FrontEnd/src/features/Dashboard/PortfolioFiles/DeletePortfolioFile.tsx
@@ -1,7 +1,7 @@
 import type React from "react";
 import { api } from "../../../ApiService/ApiBrain.ts";
 import { handleComponentError } from "../../../Helpers/ErrorHandler.ts";
-import { handleSuccess } from "../../../Helpers/Sweetalert.ts";
+import { handleConfirmation, handleSuccess } from "../../../Helpers/Sweetalert.ts";
 import { IconButton } from "@mui/material";
 import { DeleteForever } from "@mui/icons-material";
 
@@ -13,6 +13,8 @@ type Props = {
 const DeletePortfolioFile = ({ id, setTrigger }: Props) => {
   const handleDeleteOneFile: () => Promise<void> = async () => {
     try {
+      const confirmation = await handleConfirmation("Delete File?" , "This action can't be undone!")
+      if (! confirmation ) return
       await api.delete(`/api/files/${id}`);
       setTrigger && setTrigger((state) => !state);
       handleSuccess("Message", "File has been deleted");
