+++ FrontEnd/src/features/Dashboard/Technologies/DeleteTechnology.tsx
@@ -4,6 +4,7 @@ import type React from "react"
 import DeleteIcon from "@mui/icons-material/Delete";
 import { handleComponentError } from "../../../Helpers/ErrorHandler.ts";
 import { api } from "../../../ApiService/ApiBrain.ts";
+import { handleConfirmation } from "../../../Helpers/Sweetalert.ts";
 
 type Props = {
   setTrigg: React.Dispatch<SetStateAction<boolean>>
@@ -14,6 +15,9 @@ function DeleteTechnology({setTrigg , id}: Props) {
   
   const handleDeleteTechnology : ()=>Promise<void> = async ()=>{
     try {
+
+      const confirmation = await handleConfirmation("Delete Technology?","This will be permantly deleted")
+      if (!confirmation) return ;
       await api.delete(`/api/technologies/${id}`)
       setTrigg(state=>!state)
     } catch (error) {
